Test commit. - Adds world2.slime (iirc this is the one with a zombie entity.) - Fixed bitset set. - Add zombie to chunk 0,0 when player joins. - SlimeProvider is almost complete just need to load chunks/generate SlimeFile if null.

+++ src/JaxkDev/SlimeWorld/BitSet.php
@@ -64,7 +64,7 @@ class BitSet{
 	}
 
 	public function set(int $value, bool $state = true): void{
-		$part = floor($value/self::PART_SIZE);
+		$part = (int)floor($value/self::PART_SIZE);
 		$index = $value%self::PART_SIZE;
 		if(!array_key_exists($part, $this->bitset)){
 			$this->bitset[$part] = 0;
